pgsql: perform delete of replication slots during start in replication mode
This fix an issue when the dead old master keeps the replication slots on restart (if cleanup was not done manually)

The initialization of slots on Master will be done during promotion.
As stated by pacemaker documentation:
"when the resource is started, it must come up in the mode called slave. From there the cluster will decide which instances to promote to master"
this guarantees that the Master always has the slots available when slave nodes try to connect to it

(cherry picked from commit 7052fcf0641d42f67295185958fabf79cf8a8fdd)